+++ nn/modules.py
@@ -81,8 +81,8 @@ class Linear(BaseModule):
         """
         super(Linear, self).__init__(**kwargs)
         # 权重参数
-        weight = np.random.randn(in_units, out_units).astype(np.float) * np.sqrt(2 / in_units / out_units)
-        bias = np.zeros(out_units).astype(np.float)
+        weight = np.random.randn(in_units, out_units).astype(np.float32) * np.sqrt(2 / in_units / out_units)
+        bias = np.zeros(out_units).astype(np.float32)
         # 权重对应的梯度
         g_weight = np.zeros_like(weight)
         g_bias = np.zeros_like(bias)
@@ -156,8 +156,8 @@ class Conv2D(BaseModule):
         self.padding = padding
         self.stride = stride
         # 权重参数
-        weight = np.random.randn(in_filters, out_filters, *kernel).astype(np.float) * np.sqrt(2 / in_filters)
-        bias = np.zeros(out_filters).astype(np.float)
+        weight = np.random.randn(in_filters, out_filters, *kernel).astype(np.float32) * np.sqrt(2 / in_filters)
+        bias = np.zeros(out_filters).astype(np.float32)
         # 梯度
         g_weight = np.zeros_like(weight)
         g_bias = np.zeros_like(bias)
